VENV_Tezat/Tezat/requirements.txt
Django==1.11
Pillow==5.1.0
django-cleanup==3.2.0
django-nocaptcha-recaptcha==0.0.2
django-ckeditor==5.7.1
djangorestframework==3.8.2
django-countries==5.3.3
psycopg2==2.8.3
django-notifications-hq==1.5.0
